Get rid of the PERL_CORE hacks
In blead, now that we've improved how we're building extensions, we no
longer reply on modules' Makefile.PLs to specify an empty MAN3PODS.
Hence I've been going through ext/ working out what could be simplified.
Would it be possible to apply the appended patch to Digest::MD5 and
MIME::Base64, so that next time you upload it to CPAN (no rush), we can
incorporate it into blead?

Nicholas Clark