quoted-print.t: Generalize for non-ASCII
This had hard-coded just one EBCDIC code page, and would not work on
others.  The ASCII and EBCDIC code was disjoint, and over the years,
changes had been made to the ASCII version, but not the EBCDIC.

This commit joins the disparate paths so that they will stay in sync,
and modernizes to use the is() test function, which it defines itself
rather than adding dependencies.

Because the diff listing is so disjointed, I manually checked that the
final result included the tests from the ASCII version, which had more
tests than the EBCDIC, and some revisions.

The one change is that I changed "\x20" occurrences to " ".  I know
enough about perl parsing to know that the \x version will be
transformed into the literal blank before the module sees it, so that
keeping the \x version doesn't actually test anything.